chore(move): useDelayedFlag to composables
