+++ src/components/Editor.vue
@@ -98,6 +98,7 @@ import { t } from '@nextcloud/l10n'
 import { generateRemoteUrl } from '@nextcloud/router'
 import { Awareness } from 'y-protocols/awareness.js'
 import { provideConnection } from '../composables/useConnection.ts'
+import { useDelayedFlag } from '../composables/useDelayedFlag.ts'
 import { useEditorMethods } from '../composables/useEditorMethods.ts'
 import { provideEditorWidth } from '../composables/useEditorWidth.ts'
 import { provideSaveService } from '../composables/useSaveService.ts'
@@ -125,7 +126,6 @@ import ContentContainer from './Editor/ContentContainer.vue'
 import DocumentStatus from './Editor/DocumentStatus.vue'
 import MainContainer from './Editor/MainContainer.vue'
 import Status from './Editor/Status.vue'
-import { useDelayedFlag } from './Editor/useDelayedFlag.ts'
 import Wrapper from './Editor/Wrapper.vue'
 import MenuBar from './Menu/MenuBar.vue'
 import Translate from './Modal/Translate.vue'
